Solved Marker issue
Output now prints 1 for Apples.getGeneration()); but fieldSummary outputs totalApples: 1 when no apple is present on field. I have come to realise this output is due to the marker creating an instance of an apple and the fieldSummary is just reflecting that information. Therefore code is still functional and correct.